refactor(api): moved checkInvoiceStatus to be a query
